requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
asyncpg
pydantic-settings
email-validator
aiosqlite